agrego red dinamica

+++ validar-echo-server.sh
@@ -2,12 +2,18 @@
 
 
 
-NETWORK=testing_net
+
 SERVER_CONTAINER=server
 PORT=12345
 MESSAGE="ping"
 
+NETWORK=$(docker inspect $SERVER_CONTAINER \
+    --format '{{range $k, $v := .NetworkSettings.Networks}}{{$k}}{{end}}' 2>/dev/null)
 
+if [ -z "$NETWORK" ]; then
+    echo "action: test_echo_server | result: fail"
+    exit 0
+fi
 
 MSG_RESPONSE=$(docker run --rm --network $NETWORK \
     busybox sh -c  "echo $MESSAGE | nc $SERVER_CONTAINER $PORT" | tr -d '\r\n')
